blossom: widen blossom_blobs.size to BigInteger (was 32-bit; >2.1 GB blobs failed)
A blob larger than ~2.1 GB (INT4 max) failed to insert with psycopg2 NumericValueOutOfRange —
the size column was Integer while the sibling table already used BigInteger. This also wedged
stream-VOD finalize in a 30s retry loop (a 2.57 GB recording could never write its row).

Model → BigInteger, plus an idempotent type-widen in _run_migrations (ALTER COLUMN size TYPE
BIGINT, skipped once already big). On the next restart the migration widens the column and the
stuck VOD's finalize insert succeeds, so the retry loop clears itself.

Not deployed (holding — user is mid-upload; needs a restart).

+++ app/models.py
@@ -253,7 +253,7 @@ class BlossomBlob(Base):
 
     sha256 = Column(String(64), primary_key=True)
     pubkey = Column(String(64), nullable=False)        # owning uploader (hex x-only)
-    size = Column(Integer, nullable=False)
+    size = Column(BigInteger, nullable=False)   # bytes — BigInteger: a >2.1 GB blob overflowed INT4
     mime = Column(String(120), nullable=True)
     created_at = Column(Integer, nullable=False)        # unix seconds
     expires_at = Column(Integer, nullable=True)         # unix seconds; 0/NULL = never
